Fix reply links not attaching inside some tables
To fix the issue of reply links appearing inside archived discussions, I
made the code remove all reply links inside any tables. This is a bad
idea if the whole talk page is wrapped in a table, so I fixed the code
that did the removal so that it ignored tables outside of the "main
content element".